requirements.txt
paho-mqtt>=1.6.1
firebase-admin>=6.2.0
